Codigo rodado nas tres rodadas classificatorias

+++ TUR/ErusBot/PIDControl.h
@@ -1,7 +1,7 @@
 #ifndef PIDCONTROL_H
 #define PIDCONTROL_H
 
-void setupPIDControl(int kpp, int kip, int kdp);
+void setupPIDControl(double kpp, double kip, double kdp);
 int getPIDControl(int erro);
 
 #endif
